+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Changed
+
+- Sensitive environment variables (KEY, TOKEN, SECRET, PASSWORD, etc.) are now masked in `sbox info` and `sbox env list` output
+
 ## v1.3.1
 
 ### Added